Added fallback for retrieving book_id using title+author
Filename changes can generally be recovered: a new file table entry is
added.

Cover and content changes affect the file hash, upon which the PB reader
may create a new book entry. This can be 'recovered' by matching on the
filename/path. On failure, both title and author are now used as a
fallback.

Notes:
- Title changes -> ideally use lower case
- check also on year
- remove temporary addition to installed_books_by_title